refactor: 회고 - 집중력 컴포넌트 분리
