+++ src/app/(protected)/retrospection/[taskId]/RetrospectionPageClient.tsx
@@ -7,25 +7,10 @@ import { TaskResponse } from "@/types/task";
 import { useRouter } from "next/navigation";
 import RetrospectItem from "./_components/RetrospectItem";
 import { useRef, useState } from "react";
+import RetrospectFocusContent from "./_components/RetrospectFocusContent";
 
 type Props = {
     task: TaskResponse;
-}
-
-type RetrospectItems = {
-    [key: string]: {
-      title: string;
-      required: boolean;
-    };
-};
-
-type ResultContent = -1 | 0 | 1 | 2 | 3 | 4;
-type FocusContent =  0 | 1 | 2 | 3 | 4 | 5;
-
-type RetrospectContent = {
-    result: ResultContent;
-    focus: FocusContent;
-    keepAndTry?: string;
 }
 
 export default function RetrospectionPageClient({ task }: Props) {
@@ -52,54 +37,6 @@ export default function RetrospectionPageClient({ task }: Props) {
     const hasSelectedResult = () => {
         return retrospectContent.result !== -1;
     }
-
-    const setFocusContent = (selected: number) => {
-        const selectedFocus = selected as FocusContent;
-        setRetrospectContent((prev) => ({
-            ...prev,
-            focus: selectedFocus,
-        }));
-    };
-
-    const getClosestIndex = (x: number): number => {
-        console.log(`getClosestIndex: ${x}`);
-        const track = trackRef.current;
-        if (!track) return retrospectContent.focus;
-      
-        const MAX_INDEX = FOCUS_STEPS.length - 1;
-
-        const rect = track.getBoundingClientRect();
-        const offsetX = x - rect.left;
-        const ratio = offsetX / rect.width;
-        const rawIndex = Math.round(ratio * MAX_INDEX);
-        console.log(`rawIndex: ${rawIndex}`);
-        return Math.min(Math.max(rawIndex, 0), MAX_INDEX);
-    };
-
-    const handleMouseMove = (e: MouseEvent) => {
-        console.log(`handleMouseMove: ${e.clientX}`);
-        if (!isDragging.current) return;
-        const idx = getClosestIndex(e.clientX);
-        setFocusContent(idx);
-    };
-      
-    const handleMouseDown = (e: React.MouseEvent) => {
-        console.log(`handleMouseDown: ${e.clientX}`);
-        const idx = getClosestIndex(e.clientX);
-        console.log(`handleMouseDown Result: ${idx}`);
-        setFocusContent(idx);
-        isDragging.current = true;
-        
-        window.addEventListener("mousemove", handleMouseMove);
-        window.addEventListener("mouseup", handleMouseUp);
-    };
-      
-    const handleMouseUp = () => {
-        console.log("handleMouseUp");
-        isDragging.current = false;
-        window.removeEventListener("mousemove", handleMouseMove);
-        window.removeEventListener("mouseup", handleMouseUp);
-    };
 
     const hasRequiredContent = hasSelectedResult()
         && retrospectContent.focus !== undefined;
@@ -171,66 +108,10 @@ export default function RetrospectionPageClient({ task }: Props) {
 
                         {/* 몰입하는 동안 나의 집중력 */}
                         <RetrospectItem title={retrospectItems.focus.title} required={retrospectItems.focus.required}>
-                            <div className="w-full mx-2">
-                                <div 
-                                    ref={trackRef}
-                                    className="relative h-6 flex items-center"
-                                    onMouseDown={handleMouseDown}
-                                >
-                                    {/* 전체 바 배경 */}
-                                    <div 
-                                        className="absolute h-6 rounded-full bg-gray-600"
-                                        style={{
-                                            width: "calc(100% + 24px)", // 16px 양쪽 추가
-                                            left: "-12px",              // 왼쪽으로 16px 이동
-                                        }}
-                                    ></div>
-
-                                    {/* 선택된 채워진 부분 */}
-                                    <div
-                                    className="absolute h-6 rounded-full bg-gradient-to-r from-blue-200 to-purple-200 transition-all duration-200"
-                                    style={{
-                                        width: `calc(${(retrospectContent.focus / 5) * 100}% + 24px)`,
-                                        left: `-12px`,
-                                    }}
-                                    ></div>
-
-                                    {/* 점들 */}
-                                    <div className="relative z-10 flex justify-between w-full">
-                                        {FOCUS_STEPS.map((step, i) => (
-                                            <div
-                                            key={i}
-                                            className={`w-[6px] h-[6px] rounded-full transition-all duration-200 ${
-                                                retrospectContent.focus >= step
-                                                ? "bg-background-skyblue opacity-90"
-                                                : "bg-background-skyblue opacity-30"
-                                            }`}
-                                            onClick={() => setFocusContent(i)}
-                                            />
-                                        ))}
-                                    </div>
-
-                                    {/* 슬라이더 핸들 */}
-                                    <div
-                                    className="absolute m-3 z-20 w-6 h-6 rounded-full border-2 border-white bg-white shadow"
-                                    style={{
-                                        left: `calc(${(retrospectContent.focus / 5) * 100}% - 24px)`,
-                                        transition: "left 0.2s ease",
-                                    }}
-                                    />
-                                    </div>
-
-                                {/* 아래 숫자 레이블 */}
-                                <div className="mt-1 flex justify-between text-gray-400 text-sm font-medium">
-                                    {FOCUS_STEPS.map((step, i) => (
-                                        <div key={i} className="w-[6px] flex justify-center">
-                                            <span key={i} className={retrospectContent.focus === step ? "text-white" : ""}>
-                                                {`${step*20}`}
-                                            </span>
-                                        </div>
-                                    ))}
-                                </div>
-                            </div>
+                            <RetrospectFocusContent 
+                                retrospectContent={retrospectContent}
+                                setRetrospectContent={setRetrospectContent}
+                            />
                         </RetrospectItem>
 
                         {/* 몰입 회고 텍스트 */}
